Chapter 1 - Searching a Collection/List
In Chapter, we're code with a bunch of variables. We need to search the variables for two different items.

Because of the nature of the variables, one block should be turned into a simple list of stars, so we can return a value by its index (the 4th closest star to Earth = index 3). And the other block should be rewritten as a dictionary, so we can return a value by its key (variable values are described by their variable names).

Then we need to print the values we search for.

The following JupyterLab Notebook file contains the Code Stub + steps taken to reach a final code version to solve this problem.