Coding Challenge - for loop on a list
The code stub provided has a list of fruits. The goal is to return a printed "list" called "our fruit selection:" with each fruit from the list on its own line.

The provide Jupyter Notebook fall walks step-by-step through the process of writing the initial print statement, printing the list, adding a for loop, & printing the item variable within the list with each iteration.